Rename logo_dark.png to logo_darkOLD.png
we are updating the cuphead html 5 boot up screen to match with the game and website alittle more so the old one was renamed to "logo_darkOLD.png" rather than "logo_dark.png" and a new png file has taken its place if you are a user suffering the lack of being able to play right now then not to worry! this little update will finish in just 5 minutes or way less even! thanks for understanding - FFR :D...